+++ src/features/edit-user-profile/index.ts
@@ -1,1 +1,3 @@
-export { default as UserProfileCard } from './ui/UserProfileCard';
+export { default as usePatchUserProfileMutation } from './model/hooks/usePatchUserProfileMutation';
+export { default as ProfileImageUploadField } from './ui/ProfileImageUploadField';
+export type { default as UpdatedUserProfile } from './model/types';
